Light linking: initial DNA and RNA implementation
The light linking is defined by a collection which contains objects
which are receiving light. The collection itself is stored on a light
emitters.

In order to support both lamps and mesh lights the collection is
stored on the Object level.

The collection is referenced: as in, it is not one of those embedded
collections.

The dependency graph needs more work, as it currently designed to
use collections as an instancing entity. So, some refactor is needed.

Ref #104972