Depsgraph: Add explicit object collection hierarchy tracking
The practical problem this change fixes is the Cycles render
re-set on object selection.

The described behavior was introduced in the initial light
linking evaluation implementation which was only possible in the
same manner as the base flags evaluation. This is because the
outliner does not provide explicit tag when object or collection
is moved across collections.

This change limits the number of objects which are updating
their receiver mask and are updated on the Cycles side.

Pull Request: https://projects.blender.org/blender/blender/pulls/105837