Light linking: Implement drop on the receiver collection list
It is possible to drop objects and collections onto the receiver collection
list in the shading panel of object to add them to the collection.

Is not as useful yet as it could be: we need to make the list to have some
minimum height to make it possible to drop on the empty collection.

Ref #104972

Pull Request: https://projects.blender.org/blender/blender/pulls/106101