Light linking: Interface for inclusion/exclusion, as well as shadow linking
The membership in the light linking collection now has two extra options:
light state and shadow state. They define the relation of the emitter such
as inclusion/inclusion to the light reception, and same thing for shadow
casting.

The masks calculated in the dependency graph needs to be adjusted to use
the new logic. Currently they still only consider the membership in the
light linking collection.

Ref #104972