Light linking: Implement light exclusion
Allows to easily exclude specific objects from specific lights by linking
the object to the light ans setting Light State to Exclude.

The dependency graph builder is reworked once more and now it does more
explicit tracking of receiver-centric view.

The downside of this is that it kind of cancels out the memory optimization
from previous change in the builder from Brecht.

The upside is that the code is much easier to follow.

The mitigation for the increased memory consumption during the build is to
clear the receiver-centric view after the light linking has finished building.
This makes it so the overall memory footprint of the dependency graph does not
change that much at all. This is because the extra memory used by the light
linking builder is freed before the relations builder. This means that there
are no relations crated yet, and no temporary memory is used by the relation
builder (and it could be a lot!). So from the overall dependency graph build
process this change is very unlikely to become a problem.

Note that this does not mean optimizations would not happen in the future,
but it will be easier to do them once the shadow linking is laid down, and
thee are automated regression tests to help catching possible issues.

Alternatively, depending on a desired UI representations and the USD
integration we might decide that having the receiver view cached is good
to have.

Ref #104972