Light linking: Optimize memory usage during build
The idea is to replace the actual sets with corresponding
collection bitmasks during build.

The code generalizes some accessors and moves code around
to hopefully help readability. It fees like some of the
concepts will be needed for the shadow linking, but deeper
generalization is better to happen once we start working
on it.

Ref #104972