Light linking: Use more deterministic linking order
This change fixes confusion when a collection is added to
the light linking collection and is excluded, and an object
from that collection is added to the same light linking
collection and is included.

The mental model is: the more detailed object information
wins. IN a case of a tie the latest specification wins.

The current limitation is that it is very hard to re-order
the light linking collection in the list representation.
This will be worked on further after the underlying supports
hinting about items being dropped in-between of elements.

Ref #104972

Pull Request: https://projects.blender.org/blender/blender/pulls/106389