Fix shadow linking on GPU in complex scenes
The memory saving trick to re-use intersection array to store the ray self
primitives does not work reliably when there are more than 4 transparent
intersections: second time the shadow intersection kernel is invoked the
self primitives are assigned to the intersection result from the previous
kernel execution.

This did not manifest as an issue to artists before because the indices
were from the corresponding entities: self light object was stored in the
object index, and the light primitive was stored in the primitive index.
This would lead to a incorrect render result (worst case), but will not
lead to the off-the-bounds array access.

With the light ID the same trick does lead to the off-the-bounds access
as there is likely to be more objects than lights in the scene.

Likely there was unused `object` field in the `shadow_ray` of the shadow
path state, which seemed to have been always allocated. This field is now
renamed to `self_light` and the light index for the shadow linking is stored
there. This makes it possible to have reliable way of accessing light data
from the intersection kernel, as well as should lower the memory usage for
the scenes where no light linking is used.

It is not really clear what is the best way of solving the rest of the
shadow ray self primitives issue. Perhaps, save more memory in the state
and store those indices explicitly.

Even despite an extra condition added the performance is within a noise
floor different in the benchmark scenes.