Fix crash on inserting links on simulation zone nodes.
When adding links to the sim zone nodes they modify the tosock/fromsock
with the socket generated for the new item. This search expects a socket
identifier rather than a name string. The identifier is now unique, so
that sockets can be reliably relinked, but the old code was using the
name instead.

Pull Request: https://projects.blender.org/blender/blender/pulls/107303