Light linking: Simplify toggling receiver/blocker inclusion
It seems more clear to store the inclusion flag as an enumeration
internally, but on the interface it is more convenient to use a
single click to toggle the state.

With a bit of code the best of both worlds is achieved!

Pull Request: https://projects.blender.org/blender/blender/pulls/107222