Simulation Nodes: support storing "anonymous" attributes in sim state
Non-geometry sockets now work properly in simulation input/output
nodes. When a field is passed into them, it is evaluated one the
preceding geometry (this aspect might need further refinement)
and is stored as attribute. The evaluation domain can be changed
in the side bar similar to modifiers.

The attributes stored in the simulation state are not actually anonymous
but normal named attributes. This makes it much more obvious how
to the disk storage works. When the simulation state is output in
geometry nodes, the attributes become anonymous attributes again.

Anonymous attributes are not stored in the simulation state anymore,
because this can lead to unreliable and hard-to-predict results when
the set of required anonymous attributes changes over time.